split the controller actions up
