+++ src/routes/dashboardRoutes.ts
@@ -1,11 +1,14 @@
-import { showDashboard, showRevokePage, revokeAPIKey, showCreatePage, showSuccessPage} from '../controllers/dashboardController';
+import express, { Router } from 'express';
+import { showDashboard} from '../controllers/dashboardController';
+import { showCreatePage, showSuccessPage} from '../controllers/createKeyController';
+import { showRevokePage, revokeAPIKey} from '../controllers/revokeController';
 
 const router: Router = express.Router()
 
 router.get('/dashboard/:fpoId', showDashboard);
 router.get('/keys/:fpoId/:customerKeyId/revoke', showRevokePage);
 router.get('/keys/:fpoId/create', showCreatePage);
-router.post('/keys/success', showSuccessPage);
+router.post('/keys/:fpoId/success', showSuccessPage);
 router.post('/keys/:fpoId/:customerKeyId/revoke', revokeAPIKey);
 
 export default router
